refactor: Split test_openmeteo_client.py into focused modules
BEFORE/AFTER METRICS:
- Before: 501 lines in single test_openmeteo_client.py file
- After: 526 lines across 6 organized files (5% increase due to structure)
- Reduced largest file from 501 to 153 lines (70% reduction)
- All 29 tests passing with improved maintainability

FUNCTIONAL ORGANIZATION:
- tests/openmeteo_client/conftest.py (136 lines): Shared fixtures and test data
- tests/openmeteo_client/test_api_methods.py (153 lines): API method tests
- tests/openmeteo_client/test_http_handling.py (134 lines): HTTP/network error handling
- tests/openmeteo_client/test_utilities.py (59 lines): Utility and validation tests
- tests/openmeteo_client/test_client_initialization.py (43 lines): Client init tests

KEY CHANGES:
- Extracted large sample test data constants to shared conftest.py
- Organized tests by functionality rather than chronological order
- Maintained all existing test functionality and isolation
- Followed established refactoring patterns from weather_service module
- Each module stays well under 300-line maintainability threshold

BACKWARD COMPATIBILITY:
- All 29 original tests preserved and passing
- No changes to test logic or assertions
- Maintained proper imports and fixture usage
- Test coverage remains identical

TASK REFERENCE: Continuing test file refactoring work following successful
pattern established for test_weather_service.py and test_openmeteo_integration.py
