+++ tests/openmeteo_client/__init__.py
@@ -1,0 +1,1 @@
+"""OpenMeteo client tests package."""
